Restructuring services to different files. Renamings. Refactoring.
